pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "splitpost"
version = "0.1.0"
description = "One post → every platform. AI-powered content adaptation."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "aiosqlite>=0.20",
    "pydantic>=2.10",
    "pydantic-settings>=2.7",
    "pyjwt>=2.10",
    "httpx>=0.28",
]
